% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{closest}
\alias{closest}
\title{Calculate closest value in a vector to the specified value}
\usage{
closest(x, sv, na.rm = T)
}
\arguments{
\item{x}{a numeric vector}

\item{sv}{a specific value for which you want to find the closest value in a vector}
}
\value{

}
\description{
Calculate closest value in a vector to the specified value
}
\examples{
x <- c(1,2,3,4,5,6,5,4,3,2,1,99); sv = 100; closest(x, sv = sv)
}
