% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv}
\alias{cv}
\title{CV
calculates CV in the classic way assuming a normal distribution. sensitive to skewed data and outliers}
\usage{
cv(x, na.rm = T)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{default = TRUE}
}
\value{
the coefficient of variation
}
\description{
CV
calculates CV in the classic way assuming a normal distribution. sensitive to skewed data and outliers
}
\examples{
n <- 100
sd <- 0.2
u <- 0.5
set.seed(123)
y <- rnorm(n = n, mean = u, sd = sd)
cv(y)
}
