% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Estimate the mode of a numeric distribution}
\usage{
estimate_mode(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
estimated value being the maximum value in the density function of a numeric vector
}
\description{
Estimate the mode of a numeric distribution
}
\examples{
a <- c(1,2,3,4,5,5,6); estimate_mode(a)
}
