% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geo_mean}
\alias{geo_mean}
\title{Calculate the geometric mean of a numeric vector}
\usage{
geo_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{TRUE or FALSE}
}
\value{
the geometric mean
}
\description{
Calculate the geometric mean of a numeric vector
}
\examples{
a <- c(1,2,3,4,5,5,6); geo_mean(a)
}
