% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mode}
\alias{get_mode}
\title{Calculate the mode of a numeric or character vector}
\usage{
get_mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric or character vector}

\item{na.rm}{TRUE or FALSE}
}
\value{
value with the highest number of occurrences in x
}
\description{
Calculate the mode of a numeric or character vector
}
\examples{
a <- c(1,2,3,4,5,5,6); get_mode(a)
b <- c("dog","cat","dog","bird"); get_mode(b)
d <- c(1,2,3,4,NA,NA,NA,5,5,6); get_mode(d) # also works with NA
}
