% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rcvm}
\alias{rcvm}
\title{Robust CV based on MAD
As per Arachige et al 2020 arXiv:1907.01110v3}
\usage{
rcvm(x, na.rm = T, constant, ...)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{TRUE or FALSE}

\item{constant}{a constant value you supply}

\item{...}{other arguments to be passed}
}
\value{

}
\description{
Robust CV based on MAD
As per Arachige et al 2020 arXiv:1907.01110v3
}
\examples{
n <- 100
sd <- 0.2
u <- 0.5
set.seed(123)
y <- rnorm(n = n, mean = u, sd = sd)
rcvm(y)
}
