% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rcvq}
\alias{rcvq}
\title{Robust CV based on quantiles
As per Arachige et al 2020 arXiv:1907.01110v3}
\usage{
rcvq(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{default = TRUE}

\item{...}{}
}
\value{

}
\description{
Robust CV based on quantiles
As per Arachige et al 2020 arXiv:1907.01110v3
}
\examples{
n <- 100
sd <- 0.2
u <- 0.5
set.seed(123)
y <- rnorm(n = n, mean = u, sd = sd)
rcvq(y)
}
