% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rsd}
\alias{rsd}
\title{Relative standard deviation}
\usage{
rsd(x, na.rm = T)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{}
}
\value{

}
\description{
Relative standard deviation
}
\examples{
n <- 100
sd <- 0.2
u <- 0.5
set.seed(123)
y <- rnorm(n = n, mean = u, sd = sd)
rsd(y)
}
