% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upper_cl_boot}
\alias{upper_cl_boot}
\title{Bootstrapped upper confidence limit}
\usage{
upper_cl_boot(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{arguments to be passed to smean.cl.boot()}
}
\value{

}
\description{
Just a wrapper around Hmisc.cl.boot() to extract upper cl
}
\examples{
df <- data.frame(y = seq(1:10), x = c(51,52,53,54,55,55,54,53,52,99), g = c(rep("a",5),rep("b",5)))
library(tidyverse);library(magrittr)
df \%>\% group_by(g) \%>\% dplyr::summarise(u = upper_cl_boot(x,B=100))
}
